Added IntegrationTest for Game

+++ Spring/src/main/java/pl/ttsw/GameRev/service/GameService.java
@@ -97,7 +97,7 @@ public class GameService {
         return gameMapper.toDto(gameRepository.save(updatedGame));
     }
 
-    public boolean deleteGame(Integer id) {
+    public boolean deleteGame(Long id) {
         if (gameRepository.existsById(id)){
             gameRepository.deleteById(id);
             return true;
